% Auto-generated: do not edit by hand
\name{dtiIconSchoolOff}

\alias{dtiIconSchoolOff}

\title{IconSchoolOff component}

\description{

}

\usage{
dtiIconSchoolOff(id=NULL, className=NULL, color=NULL, size=NULL, stroke=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Lists containing elements .
those elements have the following types:
. Icon color}

\item{size}{Lists containing elements .
those elements have the following types:
. Icon size}

\item{stroke}{Lists containing elements .
those elements have the following types:
. Stroke width}

\item{style}{Logical | numeric | character | named list | unnamed list. Inline style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

